+++ darwinxref/plugins/exportProject.c
@@ -36,7 +36,7 @@
 #include <unistd.h>
 
 static int run(CFArrayRef argv) {
-	int res = 0;
+	ssize_t res = 0;
 	CFIndex count = CFArrayGetCount(argv);
 	if (count > 2)  return -1;
 	int xml = 0, i;
@@ -84,11 +84,11 @@ static int run(CFArrayRef argv) {
 	CFPropertyListRef plist = preplist;
 	if (xml) {
 		CFDataRef data = CFPropertyListCreateXMLData(NULL, plist);
-		res = write(STDOUT_FILENO, CFDataGetBytePtr(data), CFDataGetLength(data));
+		res = write(STDOUT_FILENO, CFDataGetBytePtr(data), (ssize_t)CFDataGetLength(data));
 	} else {
 		res = writePlist(stdout, plist, 0);
 	}
-	return res;
+	return (int)res;
 }
 
 static CFStringRef usage() {
